requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
playwright==1.41.0
aiohttp==3.9.1
Pillow==10.2.0
pydantic==2.5.3
python-multipart==0.0.6
